pyproject.toml
[project]
name = "cloudrun-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.13",
    "google-genai>=1.21.1",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.0",
]
